* 'master' of ssh://git.sip-router.org/sip-router:
  dialog(k): Note dependency between dlg_manage() and transactions in docs.
  snmpstats: control exporting registrar records
  ctl: used flavour's NAME in path of default ctl socket
  modules_k/(auth|misc)_radius: added cast
  kamailio.cfg: added storage of src ip to extra acc example
  kamailio.cfg: load cfg_rpc module to allow runtime updates
  cfg_rpc: improved documentation
  modules_k/osp, modules_k/siputils: The bind_siputils function takes 1 parameter, but the bind command in here (and as used in osp) was specified as taking 0 parameters.
  modules_k/presence_xml: The presentity activities check wasn't quite right.
  modules_k/dispatcher: Fixes to handling of empty DB and hostnames that cannot be resolved
  modules/app_lua: Fix to lua_sr_pv_is_null()
  modules:sipcapture: fixed uint types and IPPROTO_IPIP on Solaris.
  core: auto_bind_ipv6 : changed param from cfg framework to normal global parameter as setting it at run time doesn't change anything
  core: added parameter to enable/disable automatic binding of ipv6 interfaces TODO: documentation
  core:socket_info Added automatic discovery of IPv6 interfaces on Linux using RT_NETLINK socks
  dialog(k): Fix typos.
  Revert "message shooter module added"

+++ modules_k/auth_radius/extra.c
@@ -190,7 +190,7 @@ int extra2strar( struct extra_attr *extra, struct sip_msg *rq, str *val_arr)
 	    val_arr[n].len = 0;
 	} else if (value.flags&PV_VAL_INT) {
 	    /* len = -1 denotes int type */
-	    val_arr[n].s = (char *)value.ri;
+	    val_arr[n].s = (char *)(long)value.ri;
 	    val_arr[n].len = -1;
 	} else {
 	    /* set the value into the acc buffer */
